pa_backend/requirements.txt
Flask==2.1.2
Flask-Cors==3.1.2
tensorflow-cpu==2.9.1
numpy==1.23.5
Pillow==9.3.0
h5py==3.6.0